bugfix(rhineng-12147): Handle systems without owner_id to avoid spurious wildcard deletion
* Add guards for systems without owner_id
* Move code from the produce events code as requested

+++ lib/group_repository.py
@@ -60,11 +60,16 @@ def _produce_host_update_events(event_producer, host_id_list, group_id_list=[], 
         )
         event = build_event(EventType.updated, serialized_host, platform_metadata=metadata)
         event_producer.write_event(event, serialized_host["id"], headers, wait=True)
+    return host_list
 
+
+def _invalidate_system_cache(host_list):
+    identity = get_current_identity()
     for host in host_list:
         insights_id = host.canonical_facts.get("insights_id")
         owner_id = host.system_profile_facts.get("owner_id")
-        delete_cached_system_keys(insights_id=insights_id, org_id=identity.org_id, owner_id=owner_id)
+        if insights_id and owner_id:
+            delete_cached_system_keys(insights_id=insights_id, org_id=identity.org_id, owner_id=owner_id)
 
 
 def _add_hosts_to_group(group_id: str, host_id_list: List[str]):
@@ -116,7 +121,10 @@ def add_hosts_to_group(group_id: str, host_id_list: List[str], event_producer: E
         _add_hosts_to_group(group_id, host_id_list)
 
     # Produce update messages once the DB session has been closed
-    _produce_host_update_events(event_producer, host_id_list, group_id_list=[group_id], staleness=staleness)
+    host_list = _produce_host_update_events(
+        event_producer, host_id_list, group_id_list=[group_id], staleness=staleness
+    )
+    _invalidate_system_cache(host_list)
 
 
 def add_group(group_data, event_producer) -> Group:
@@ -142,7 +150,10 @@ def add_group(group_data, event_producer) -> Group:
     created_group = Group.query.filter((Group.name == group_name) & (Group.org_id == org_id)).one_or_none()
 
     # Produce update messages once the DB session has been closed
-    _produce_host_update_events(event_producer, host_id_list, group_id_list=[created_group.id], staleness=staleness)
+    host_list = _produce_host_update_events(
+        event_producer, host_id_list, group_id_list=[created_group.id], staleness=staleness
+    )
+    _invalidate_system_cache(host_list)
 
     return created_group
 
@@ -198,7 +209,8 @@ def delete_group_list(group_id_list: List[str], event_producer: EventProducer) -
                 else:
                     log_group_delete_failed(logger, group_id, get_control_rule())
 
-    _produce_host_update_events(event_producer, deleted_host_ids, [], staleness=staleness)
+    host_list = _produce_host_update_events(event_producer, deleted_host_ids, [], staleness=staleness)
+    _invalidate_system_cache(host_list)
     return deletion_count
 
 
@@ -208,7 +220,8 @@ def remove_hosts_from_group(group_id, host_id_list, event_producer):
     with session_guard(db.session):
         removed_host_ids = _remove_hosts_from_group(group_id, host_id_list)
 
-    _produce_host_update_events(event_producer, removed_host_ids, [], staleness=staleness)
+    host_list = _produce_host_update_events(event_producer, removed_host_ids, [], staleness=staleness)
+    _invalidate_system_cache(host_list)
     return len(removed_host_ids)
 
 
@@ -268,7 +281,8 @@ def patch_group(group: Group, patch_data: dict, event_producer: EventProducer):
         # If anything was updated, and the host list is not being replaced,
         # send update messages to existing hosts. Otherwise, wait until the host list is replaced
         # so we don't produce messages that will be instantly obsoleted.
-        _produce_host_update_events(event_producer, existing_host_ids, [group_id], staleness=staleness)
+        host_list = _produce_host_update_events(event_producer, existing_host_ids, [group_id], staleness=staleness)
+        _invalidate_system_cache(host_list)
     elif new_host_ids is not None:
         # If host IDs were provided, we need to update the host list.
         # First, update the modified date for the group
@@ -276,10 +290,12 @@ def patch_group(group: Group, patch_data: dict, event_producer: EventProducer):
         db.session.add(group)
 
         deleted_host_uuids = [str(host_id) for host_id in (existing_host_ids - new_host_ids)]
-        _produce_host_update_events(event_producer, deleted_host_uuids, [], staleness=staleness)
+        host_list = _produce_host_update_events(event_producer, deleted_host_uuids, [], staleness=staleness)
+        _invalidate_system_cache(host_list)
 
         added_host_uuids = [str(host_id) for host_id in new_host_ids]
-        _produce_host_update_events(event_producer, added_host_uuids, [group_id], staleness=staleness)
+        host_list = _produce_host_update_events(event_producer, added_host_uuids, [group_id], staleness=staleness)
+        _invalidate_system_cache(host_list)
 
 
 def _update_group_update_time(group_id: str):
